+++ NBB-Project-Back-Enc/NBB.Api/Repositories/IRepository.cs
@@ -1,6 +1,14 @@
-﻿namespace NBB.Api.Repository
+﻿using NBB.Api.Models;
+using System.Linq.Expressions;
+
+namespace NBB.Api.Repository
 {
-    public class IRepository
+    public interface IRepository
     {
+        void Add(Onderneming onderneming);
+        void Delete(Onderneming onderneming);
+        Onderneming Get(string ondernemingsnummer);
+        IEnumerable<Onderneming> GetAll();
+        void Update(Onderneming onderneming);
     }
 }
